+++ src/main/java/team/unstudio/udpl/core/nms/asm/AsmNmsManager.java
@@ -7,8 +7,8 @@ import org.bukkit.inventory.ItemStack;
 
 import team.unstudio.udpl.core.UDPLib;
 import team.unstudio.udpl.nms.NmsNBT;
-import team.unstudio.udpl.nms.ReflectionUtils;
 import team.unstudio.udpl.nms.inventory.NmsItemStack;
+import team.unstudio.udpl.util.ReflectionUtils;
 
 public class AsmNmsManager {
